added user game info administrative form

+++ system/application/views/admin/games/index.tpl
@@ -30,9 +30,8 @@
                 <td><a href="games/show/{$game->gameID}">{$game->name}</a></td>
                 <td>{$game->genre}</td>
                 <td>{if $game->active == 1}Active{else}Inactive{/if}</td>
-                <td><a href="games/edit/{$game->gameID}">edit</a> | <a href="games/destroy/{$game->gameID}">remove</a></td>
+                <td><a href="games/edit/{$game->gameID}">edit</a> | <a href="games/destroy/{$game->gameID}" data-confirm="Are you sure you wish to remove this game from the server?">remove</a></td>
             </tr>
-        <p></p>
     {/foreach}
         </tbody>
         <tfoot>
